feat: 新增内置智能员工与飞书消息通道

+++ app/src-tauri/src/sidecar.rs
@@ -148,7 +148,7 @@ pub fn start_sidecar(app: &AppHandle) -> Result<(), String> {
             }
             // 指数退避：2, 4, 8, 16, 32s（failures=1..5），封顶 60s。
             // failures 从 1 开始（fetch_add 返回旧值，+1 后是新计数）。
-            let exp = failures.min(5) as u32;
+            let exp = failures.min(5);
             let backoff_secs = BASE_BACKOFF_SECS
                 .saturating_mul(1u64 << exp)
                 .min(60);
